+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Changed
+- Now using a services manager instead of CQRS resource models.
+- The module now depends on the EDD Bookings services module.
 
 ## [0.1-alpha7] - 2018-08-15
 ### Fixed
@@ -33,7 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Cart now shows booking info datetimes in the client timezone.
 
 ### Added
-- Booking items in cart now have price ID association.=
+- Booking items in cart now have price ID association.
 
 ## [0.1-alpha3] - 2018-06-11
 ### Fixed